examples/new-foundational/requirements.txt
pipecat-ai[webrtc]
fastapi
uvicorn
python-dotenv
python-fasthtml